+++ lib/mongoid.rb
@@ -20,6 +20,7 @@ require "mongoid/loggable"
 require "mongoid/sessions"
 require "mongoid/document"
 require "mongoid/unit_of_work"
+require "mongoid/log_subscriber"
 
 # If we are using Rails then we will include the Mongoid railtie. This has all
 # the nifty initializers that Mongoid needs.